OTA-2743: Restructure tests in order to reuse common functionality
